# public/docs/swagger.yaml
...
openapi: "3.0.0"
info:
  description: |
    프론트엔드, 백엔드로 작업을 하다보면 데이터가 송, 수신되는데 이 과정에서 서로가 필요한 데이터를 정확히 모르기 때문에
    이에 대한 사항을 문서로 남겨놓고자 해당 문서를 작성한다
  version: 1.1.0
  title: Mill-Planning
  termsOfService: https://github.com/ece5074/Mill-Planning-Project
  contact:
    email: ece5074@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# - http
# parameter에 대한 부분을 명세
# in: path, body, header, query 의 형태를 주로 사용
# 순서는 상관없는듯?
# name: 변수명
# type: 데이터 타입의 명시
# schema $ref: 밑 부분의 model 작성을 통해 참조
paths:
...
  /api/test/{message}:
    get:
      summary: 테스트
      parameters:
        - $ref: 'parameters.yaml#/test'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/test"
        '418':
          description: "나는 커피를 끓이는것을 원치않는다 아이스아메리카노를 가져와라"

  ###############################가게######################################
  #가게 생성
  /api/store/:
    post:
      summary: "가게 생성(관리자)"
      requestBody:
        $ref: 'requestBody/store.yaml'
      
      responses:
        '201':
          description: "해당 사용자의 가게가 생성되었음을 알림"
        default:
          $ref: 'responses/error.yaml'

  #가게 검색
...
  /api/ad-store:
    get:
      summary: "소유한 가게를 조회 ( 가게 관리자 로그인 체크 )"
      responses:
        '200':
          description: "다음의 값이 배열로 전송"
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/stores"
        default:
          $ref: 'responses/error.yaml'

  ##############################유저 또는 가게 관리자 계정관련#############################

  #로그인
...
  /api/customers/reserve:
    get:
      summary: "유저 예약 현황"
      parameters:
      responses:
        '200':
          dsecription: "현재 로그인된 사용자의 예약 현황을 불러옴"
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/reservation"
        default:
          $ref: 'responses/error.yaml'
  
  #나의 이전 예약현황(유저)
...